pyproject.toml
[project]
name = "qdrant-admin-mcp"
version = "0.1.0"
description = "FastMCP server for Qdrant administration"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.14.4",
    "logfire>=4.21.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.8.0",
    "python-dotenv>=1.0.1",
]
